Secured Delete FD controller to allow only admin and owner of FD to delete

+++ investment/src/main/java/my/investment/fd/Controllers/DeleteController.java
@@ -14,7 +14,9 @@ import org.springframework.web.server.ResponseStatusException;
 
 import my.investment.fd.Classes.Role;
 import my.investment.fd.Entities.FixedDeposit;
+import my.investment.fd.Entities.User;
 import my.investment.fd.Repositories.FixedDepositRepository;
+import my.investment.fd.Security.AuthUtil;
 
 
 
@@ -39,6 +41,11 @@ public class DeleteController {
     ) {
         FixedDeposit fd = fixedDepositRepository.findById(id).orElse(null);
         if (fd == null) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Non-existent Fixed Deposit ID: " + id);
+
+        User u = AuthUtil.getCurrentUser();
+        if (u.getRole() != Role.ROLE_ADMIN && u.getId() != fd.getUser().getId()) 
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Only admin and owner can delete this Fixed Deposit");
+
         fixedDepositRepository.delete(fd);
         return ResponseEntity.ok(fd);
     }
